uint64 sequence numbers (#599)
* Initial support for uint64_t sequence numbers

* MySQL uses BIGINT UNSIGNED
Fix output digits for filestore sequence numbers

* postgresql uses NUMERIC

* Convert mysql and postgresql sequence numbers to uint64_t

* ODBC compiles with SEQNUM

* Compiles in visual studio

* ODBC tests pass

* Regenerate swig

* Ignore store directory

* Add MessageStore tests for ruby and python

* Remove obsolete runat_python.sh

* Try to load old seqnum file format

* Add comment

+++ src/C++/FieldTypes.h
@@ -908,7 +908,7 @@ typedef UtcDateOnly UTCDATEONLY;
 typedef UtcTimeOnly UTCTIMEONLY;
 typedef int NUMINGROUP;
 typedef double PERCENTAGE;
-typedef int SEQNUM;
+typedef uint64_t SEQNUM;
 typedef int TAGNUM;
 typedef int LENGTH;
 typedef std::string COUNTRY;
