+++ src/sql/mysql/messages_log_table.sql
@@ -3,7 +3,7 @@ USE quickfix;
 DROP TABLE IF EXISTS messages_log;
 
 CREATE TABLE messages_log (
-  id INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
   time DATETIME NOT NULL,
   time_milliseconds INT NOT NULL,
   beginstring CHAR(8),
@@ -17,7 +17,7 @@ CREATE TABLE messages_log (
 DROP TABLE IF EXISTS messages_backup_log;
 
 CREATE TABLE messages_backup_log (
-  id INT UNSIGNED NOT NULL AUTO_INCREMENT,
+  id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
   time DATETIME NOT NULL,
   time_milliseconds INT NOT NULL,
   beginstring CHAR(8),
