+++ src/sql/postgresql/messages_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE messages (
   sendercompid VARCHAR(64) NOT NULL,
   targetcompid VARCHAR(64) NOT NULL,
   session_qualifier VARCHAR(64) NOT NULL,
-  msgseqnum INTEGER NOT NULL, 
+  msgseqnum NUMERIC NOT NULL, 
   message TEXT NOT NULL,
   PRIMARY KEY (beginstring, sendercompid, targetcompid, session_qualifier, msgseqnum)
 );
